New facing method for NPCs with only 2 facing directions.
Many NPCs are based on "ObjPersonNew" which has a function to change the facing direction. The problem is, this function was only useful for NPCs that have four facing directions. NPCs like Papahl only has two facing directions, so it barely worked for him. So this implements a new function to update binary facing NPCs like him. Not sure if there's more, he's the only one implemented at the moment. This also makes the facing turn rate for four direction NPCs much faster as it removes much of the delay.