Change height of one way collider.
One of those things that make no sense... When it's height is 3, Link will stand on it for a brief moment. When it's 4, he won't stand on it. When it's 3.25, he won't stand on it so we'll use that. This is still low enough for the rooster to fly over it.